+++ recipes_app/lib/domain/entitis/consistency.dart
@@ -1,0 +1,6 @@
+import '../../data/models/recipe_model.dart';
+
+enum Consitency { SOLID, LIQUID }
+
+final consitencyValues =
+    EnumValues({"liquid": Consitency.LIQUID, "solid": Consitency.SOLID});
