+++ recipes_app/lib/domain/entitis/measures.dart
@@ -1,0 +1,11 @@
+import 'package:recipes_app/domain/entitis/metric.dart';
+
+class Measures {
+  Metric metric;
+  Metric us;
+
+  Measures({
+    required this.metric,
+    required this.us,
+  });
+}
