+++ recipes_app/lib/domain/entitis/metric.dart
@@ -1,0 +1,11 @@
+class Metric {
+  double amount;
+  String unitLong;
+  String unitShort;
+
+  Metric({
+    required this.amount,
+    required this.unitLong,
+    required this.unitShort,
+  });
+}
