+++ recipes_app/lib/domain/entitis/prodcut.match.dart
@@ -1,0 +1,23 @@
+class ProductMatch {
+  int id;
+  String title;
+  String description;
+  String price;
+  String imageUrl;
+  double averageRating;
+  int ratingCount;
+  double score;
+  String link;
+
+  ProductMatch({
+    required this.id,
+    required this.title,
+    required this.description,
+    required this.price,
+    required this.imageUrl,
+    required this.averageRating,
+    required this.ratingCount,
+    required this.score,
+    required this.link,
+  });
+}
