+++ recipes_app/lib/domain/entitis/wine_pairing.dart
@@ -1,0 +1,13 @@
+import 'prodcut.match.dart';
+
+class WinePairing {
+  List<String> pairedWines;
+  String pairingText;
+  List<ProductMatch> productMatches;
+
+  WinePairing({
+    required this.pairedWines,
+    required this.pairingText,
+    required this.productMatches,
+  });
+}
